% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcor.R
\name{calcor}
\alias{calcor}
\title{Calculate the correlation coefficient.}
\usage{
calcor(
  codt = codt,
  r_th = 0.6,
  p_th = 0.05,
  type = c("cor", "mine"),
  n.cores = 4,
  group = NULL
)
}
\arguments{
\item{codt}{dataframe.}

\item{r_th}{R threshold.}

\item{p_th}{p threshold.}

\item{type}{type=c("cor","mine").}

\item{n.cores}{n.cores.}

\item{group}{group(default NULL).}
}
\description{
Calculate the correlation coefficient.
}
\examples{
crores <- calcor(codt = cordt,group = NULL,type = "cor",r_th = 0.6,p_th = 0.05,n.cores = 4)
crores$modularity #get modularity
ndedinf <- get_node_edg(gg = crores,nc = FALSE) #get nodes and edges inf

}
\seealso{
\code{\link{rcorr}}, \code{\link{mine}}
}
