% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mult_sb_nt.R
\name{get_mult_sb_nt}
\alias{get_mult_sb_nt}
\title{Get subnetwork.}
\usage{
get_mult_sb_nt(ggdt = ggdt, asvdt = asvdt, delete_single_v = TRUE)
}
\arguments{
\item{ggdt}{The result of calcor.}

\item{asvdt}{dataframe.}
}
\description{
Get subnetwork.
}
