% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node_edg.R
\name{get_node_edg}
\alias{get_node_edg}
\title{Get nodes and edges inf.}
\usage{
get_node_edg(ggdt = ggdt, nc = FALSE)
}
\arguments{
\item{ggdt}{The result of calcor or net_pro_sub.}

\item{nc}{natural connectivity.}
}
\description{
Get nodes and edges inf.
}
\examples{
crores <- calcor(codt = cordt,group = NULL,type = "cor",r_th = 0.6,p_th = 0.05,n.cores = 4)
get_node_edg(ggdt = crores,nc = FALSE) ## or
ntprdt <- net_pro_sub(ggdt = crores,asvdt = as.data.frame(t(cordt)))
get_node_edg(ggdt = ntprdt,nc = FALSE) # or
ntprdt <- net_pro_sub(ggdt = crores,asvdt = as.data.frame(t(cordt)),group = group)
get_node_edg(ggdt = ntprdt,nc = FALSE)
}
