% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_single_sb_nt.R
\name{get_single_sb_nt}
\alias{get_single_sb_nt}
\title{Get subnetwork.}
\usage{
get_single_sb_nt(gg = gg, asvdt = asvdt, delete_single_v = TRUE)
}
\arguments{
\item{gg}{The object of igraph.}

\item{asvdt}{dataframe.}
}
\description{
Get subnetwork.
}
\examples{
crores <- calcor(codt = cordt,group = NULL,type = "cor",r_th = 0.6,p_th = 0.05,n.cores = 4)
sbnt <- get_single_sb_nt(ggdt = crores,asvdt = as.data.frame(t(cordt)))
}
