% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_pro_sub.R
\name{net_pro_sub}
\alias{net_pro_sub}
\title{Calculate the topology coefficients.}
\usage{
net_pro_sub(netlist = netlist)
}
\arguments{
\item{netlist}{The result of calcor.}
}
\description{
Calculate the topology coefficients.
}
\examples{
crores <- calcor(codt = cordt,group = group,type = "cor",r_th = 0.6,p_th = 0.05,n.cores = 4)
sbnt <- get_mult_sb_nt(ggdt = crores,asvdt = cordt)
ntpro <- net_pro_sub(netlist = spnt)
}
