% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntwk_per_nc.R
\name{ntwk_per_nc}
\alias{ntwk_per_nc}
\title{Calculate natural connectivity.}
\usage{
ntwk_per_nc(gg = gg, pc = 0.8, iternum = NULL, thrnm = NULL)
}
\arguments{
\item{gg}{igraph object.}

\item{pc}{Percentage.}

\item{iternum}{Number of iterations.}

\item{thrnm}{Thread.}
}
\description{
Calculate natural connectivity.
}
\examples{
ncres <- ntwk_per_nc(gg=crores$gg,pc=0.8)
ncresgrp <- ntwk_per_nc(gg=crores[[1]]$gg,pc=0.8)
}
