% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_net.R
\name{plot_net}
\alias{plot_net}
\title{plot.}
\usage{
plot_net(
  netwkinf = netwkinf,
  ex_edg = FALSE,
  grpnm_position = "ld",
  grpnm_position_num = c(1, 0.85),
  point_size = c("dg", "nc"),
  ex_sp_num = 0,
  per_modules = 0,
  nrow = 2,
  group_order = NULL,
  node_text_size = 5,
  legend_text_size = 20,
  group_text_size = 7
)
}
\arguments{
\item{netwkinf}{The results of get_node_edg.}

\item{grpnm_position}{one of "ld","lt","rd",and "rt".}
}
\description{
plot.
}
\examples{
ndedinf <- get_node_edg(gg = crores,nc = FALSE)
ptres <-  plot_net(netwkinf = ndedinf,ex_edg = F,per_modules = 0,ex_sp_num = 0,point_size = "dg",grpnm_position="ld")
}
